src/main/java/eel/seprphase2/Simulator/Persistence.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package eel.seprphase2.Simulator;

import com.fasterxml.jackson.core.JsonProcessingException;
import eel.seprphase2.Utilities.Percentage;
import com.fasterxml.jackson.databind.ObjectMapper;
/**
 *
 * @author drm511
 */
public class Persistence {

    public String serialize(Object obj) throws JsonProcessingException {
        ObjectMapper mapper = new ObjectMapper();
        return mapper.writeValueAsString(obj);
    }
    
}
